+++ backend/src/main/java/dk/pekilidi/dod/data/CharacterDTO.java
@@ -11,7 +11,6 @@ import dk.pekilidi.dod.character.model.FavoriteHand;
 import dk.pekilidi.dod.character.model.Looks;
 import dk.pekilidi.dod.character.model.SocialStatus;
 import dk.pekilidi.dod.character.model.body.BodyPartName;
-import dk.pekilidi.dod.skill.SkillKey;
 import dk.pekilidi.dod.util.rules.RulesUtil;
 import java.io.Serial;
 import java.io.Serializable;
@@ -62,7 +61,7 @@ public class CharacterDTO implements DODFact, Serializable {
   @Default
   @EqualsAndHashCode.Exclude
   @JsonInclude(Include.NON_EMPTY)
-  private Map<SkillKey, SkillDTO> skills = new HashMap<>();
+  private Map<String, SkillDTO> skills = new HashMap<>();
   private Looks looks;
 
   public Integer getBaseTraitValue(BaseTraitName baseTraitName) {
@@ -80,7 +79,7 @@ public class CharacterDTO implements DODFact, Serializable {
   }
 
   public void addSkill(SkillDTO skillDTO) {
-    skills.putIfAbsent(skillDTO.getKey(), skillDTO);
+    skills.putIfAbsent(skillDTO.getKey().getValue(), skillDTO);
   }
 
   public void addBaseTrait(BaseTraitDTO baseTrait) {