+++ backend/src/test/java/dk/pekilidi/dod/changerequest/ChangeRequestServiceTest.java
@@ -17,6 +17,8 @@ import dk.pekilidi.dod.skill.SkillKey;
 import dk.pekilidi.dod.skill.model.Group;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -97,17 +99,20 @@ class ChangeRequestServiceTest {
     assertThat(changeRequest.getStatusLabel()).isEqualTo(ChangeStatusLabel.INSUFFICIENT_HERO_POINTS);
   }
 
-  @Test
-  void buyNewSkill() {
+  @ParameterizedTest
+  @CsvSource({
+      "'char1', 'primary.weapon'",
+      "'char2', 'secondary.weapon'",
+      "'char3', 'skill.that.does.not.exist'",
+      "'char4', 'primary.weapon'"})
+  void buyNewSkill(String charName, String skillName) {
     CharacterDTO newChar = charService.createCharacter(
-        CharacterDTO.builder().name("test-char").hero(true).race(new RaceDTO("human")).build());
-    newChar.setState(INIT_COMPLETE);
-    newChar = charService.save(newChar);
+        CharacterDTO.builder().name(charName).hero(true).race(new RaceDTO("human")).build());
 
     SkillDTO skill = SkillDTO
         .builder()
         .group(Group.COMBAT)
-        .key(SkillKey.builder().value("primary.weapon").build())
+        .key(SkillKey.builder().value(skillName).build())
         .build();
 
     ChangeRequest change = ChangeRequest
@@ -119,8 +124,44 @@ class ChangeRequestServiceTest {
 
     ChangeRequest changeRequest = changeRequestService.submitChangeRequest(newChar.getId(), change);
 
+    if(skillName.equals("skill.that.does.not.exist")){
+      assertThat(changeRequest.getStatus()).isEqualTo(ChangeStatus.REJECTED);
+      assertThat(changeRequest.getStatusLabel()).isEqualTo(ChangeStatusLabel.SKILL_DOES_NOT_EXIST);
+      assertThat(changeRequest.getObjectAfterChange()).isNull();
+    } else {
+      assertThat(changeRequest.getStatus()).isEqualTo(ChangeStatus.APPROVED);
+      assertThat(changeRequest.getStatusLabel()).isEqualTo(ChangeStatusLabel.OK_SKILL_BOUGHT);
+      assertThat(changeRequest.getObjectAfterChange()).isNotNull();
+    }
+  }
+
+  @Test
+  void buyDuplicateSkillShouldFail() {
+    CharacterDTO newChar = charService.createCharacter(
+        CharacterDTO.builder().name("test-char-duplicate").hero(true).race(new RaceDTO("human")).build());
+
+    SkillDTO skill = SkillDTO
+        .builder()
+        .group(Group.COMBAT)
+        .key(SkillKey.builder().value("primary.weapon").build())
+        .build();
+
+    ChangeRequest change = ChangeRequest
+        .builder()
+        .changeType(ChangeType.NEW_SKILL)
+        .changeKey(skill.getKey())
+        .modifier(11)
+        .build();
+
+    ChangeRequest changeRequest = changeRequestService.submitChangeRequest(newChar.getId(), change);
+
     assertThat(changeRequest.getStatus()).isEqualTo(ChangeStatus.APPROVED);
     assertThat(changeRequest.getStatusLabel()).isEqualTo(ChangeStatusLabel.OK_SKILL_BOUGHT);
     assertThat(changeRequest.getObjectAfterChange()).isNotNull();
+
+    changeRequest = changeRequestService.submitChangeRequest(newChar.getId(), change);
+    assertThat(changeRequest.getStatusLabel()).isEqualTo(ChangeStatusLabel.SKILL_ALREADY_BOUGHT);
   }
+
+
 }